refraktoryzacja kodu -> rozbicie snackbara na osobny kompontent

+++ src/views/recipes/Recipe.vue
@@ -27,26 +27,7 @@
 					<div class="d-flex">
 						<v-icon class="" color="yellow darken-1" :size="20" v-for="(recipe, index) in rating" :key="recipe.rating">mdi-star</v-icon>						
 						<v-icon class="" color="yellow darken-1" :size="20" v-for="(recipe,index) in 5-rating" :key="recipe.rating">mdi-star-outline</v-icon>
-						<v-dialog v-model="dialog" persistent max-width="400">
-							<template v-slot:activator="{ on }">
-								<v-btn v-if="user.docId" @click="selectedStar = null" v-on="on" color="teal" depressed class="ml-2 d-flex white--text align-center">
-									<v-icon class="" left>mdi-star</v-icon>
-									<p class="ma-0">Oceń</p>
-								</v-btn>
-							</template>
-							<v-card class="">
-								<v-card-title class="headline teal white--text" color="teal" teal primary-title>Oceń przepis!</v-card-title>
-								<v-card-text class="pa-5 justify-center d-flex">
-
-									<v-icon class="recipe-star" color="yellow darken-1" :size="star.size" v-for="(star, index) in starType" @mouseleave="resetStars" @mouseenter="fillStars(index)" @click="selectStar(index)"  :key="index">{{star.name}}</v-icon>
-
-								</v-card-text>
-								<v-divider></v-divider>
-								<v-card-actions>
-									<v-btn color="teal" class="white--text" @click="rateRecipe">Wyślij</v-btn>
-									<v-btn color="error" class="white--text" @click="closeDialog">Zamknij</v-btn>
-								</v-card-actions></v-card>
-							</v-dialog>
+						<Rating :user="user" :recipe="recipe"/>
 						</div>
 						<div>
 							<p>Autor: {{recipeUserName}}</p>
@@ -65,34 +46,26 @@
 					<h2 class="title">Przepis:</h2>
 					<p>{{recipe.instructions}}</p>
 				</v-row>
-				<v-snackbar v-model="snackbar" :timeout="4000" class="white--text" color="teal" right>
-					{{ alertText  }}
-					<v-btn color="white" text @click="snackbar = false">
-						Zamknij
-					</v-btn>
-				</v-snackbar>
+				<Snackbar :snackbar="snackbar" :alertText="alertText" @closeSnackbar="closeSnackbar" :snackbarColor="snackbarColor" />
 			</v-container>
 		</template>
 
 		<script>
+			import Rating from '@/components/recipes/Rating';
+			import Snackbar from '@/components/Snackbar';
+
 			export default{
 				data(){
 					return{
 						snackbar: false,
 						alertText: '',
 						saved: false,
+						snackbarColor: 'teal',
 						currentUserID: localStorage.getItem('docId'),
-						dialog: false,
-						selectedStar: null,
-						starType: [
-						{name: 'mdi-star-outline', size: 50}, 
-						{name: 'mdi-star-outline', size: 50},
-						{name: 'mdi-star-outline', size: 50},
-						{name: 'mdi-star-outline', size: 50},
-						{name: 'mdi-star-outline', size: 50}
-						],
-
 					}
+				},
+				components:{
+					Rating, Snackbar
 				},
 				watch:{
 					'$route.params.id'(){
@@ -126,7 +99,13 @@
 					this.$store.commit('setCurrentRecipe', this.$route.params.id);
 				},
 				methods:{
+					closeSnackbar(){
+						console.log($event);
+					},
 					deleteRecipe(){
+						if(!confirm('Ta operacja trwale usunie Twój przepis, czy kontynuować?')){
+							return;
+						}
 						this.$store.commit('setDeleteLoading', true);
 						this.$store.dispatch('deleteRecipe',this.recipe.id);
 						this.snackbar = true;
@@ -167,74 +146,6 @@
 					editRecipe(){
 						this.$router.push('/recipeform/'+this.recipe.id);
 					},
-					rateRecipe(){
-						if(this.selectedStar == null){
-							return;
-						}
-						if(!Array.isArray(this.recipe.rating)){
-							this.recipe.rating = [];
-						}
-						this.recipe.rating.push(this.selectedStar);
-						this.$store.dispatch('updateRecipe', this.recipe);
-						for(let i =0 ;i< 5 ;i++){
-							this.starType[i].name = 'mdi-star-outline';
-							this.starType[i].size = 50;
-						}
-						this.dialog = false;
-					},
-					fillStars(index){
-						for(let i=0 ; i<index+1 ;i++){
-							this.starType[i].name = 'mdi-star';
-						}
-					},
-					closeDialog(){
-						this.dialog = false;
-						for(let i =0 ;i< 5 ;i++){
-							this.starType[i].name = 'mdi-star-outline';
-							this.starType[i].size = 50;
-						}
-					},
-					resetStars(){
-						for(let i=0 ; i<5 ;i++){
-							if(this.selectedStar == null || this.selectedStar <= i){
-								this.starType[i].name = 'mdi-star-outline';
-							}
-						}
-					},
-					selectStar(index){
-						var isSelected = false;
-						if(this.selectedStar == null){
-							this.selectedStar = index+1;
-						}else if(this.selectedStar != index+1 ){
-							this.selectedStar = index+1;
-						}else{
-							if(this.selectedStar == index+1){
-								isSelected = true;
-							}
-						}
-						for(let i=0 ; i<5 ;i++){
-							if(i == index){
-								this.starType[i].size = 80;
-								this.starType[i].name = 'mdi-star';
-
-							}else{
-								this.starType[i].size = 50;
-								this.starType[i].name = 'mdi-star-outline';
-							}
-							if(i <= index){
-								this.starType[i].name = 'mdi-star';
-							}else{
-								this.starType[i].name = 'mdi-star-outline';
-							}
-							if(isSelected){
-								this.starType[i].name = 'mdi-star-outline';
-								this.starType[i].size = 50;
-								this.selectedStar = null;
-							}
-
-						}
-						
-					}
 				}
 			}
 		</script>
@@ -247,7 +158,5 @@
 				max-width: 80%;
 				height: auto;
 			}
-			.recipe-star{
-				cursor: pointer;
-			}	
+				
 		</style>
